generated-datasets/workspace-python-api/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
